Add static cast

+++ src/components/RectangleComponent.cpp
@@ -4,10 +4,10 @@
 
 void LNComponents::RectangleComponent::render() {
   SDL_Rect sdlRect;
-  sdlRect.w = rect.size.x;
-  sdlRect.h = rect.size.y;
-  sdlRect.x = rect.position.x;
-  sdlRect.y = rect.position.y;
+  sdlRect.w = static_cast<int>(rect.size.x);
+  sdlRect.h = static_cast<int>(rect.size.y);
+  sdlRect.x = static_cast<int>(rect.position.x);
+  sdlRect.y = static_cast<int>(rect.position.y);
 
   auto canvas2D = core->getCanvas();
   canvas2D->fillRect(&sdlRect, canvas2D->getColor(RectangleComponent::color));
